Move IJP provenance call into cat_utils library
Fixes #8 .
The example provenance jobscript now creates a provenance record in
ICAT, using the storeProvenance() function from cat_utils. The jobscript
passes the ijpUrl and ijpJobId in the call, which then does the work of
informing the IJP.
The jobscript now creates a new (empty) dataset, whose name is given by
the user, and creates a provenance record linking it to the input
dataset.